eu n sei mais o que fazer da vida.......

+++ src/main/java/com/team1/investsim/dtos/AssetDTO.java
@@ -1,3 +1,5 @@
 package com.team1.investsim.dtos;
 
-public record AssetDTO(long id, String ticker, double currentPrice, double dailyChange) {}
+import java.math.BigDecimal;
+
+public record AssetDTO(long id, String ticker, BigDecimal currentPrice, BigDecimal dailyChange) {}
